Add tests for buildAspects.nested {} configuration
