+++ src/main/java/be/vbgn/gradle/buildaspects/settings/dsl/BuildAspectsImpl.java
@@ -9,13 +9,11 @@ import be.vbgn.gradle.buildaspects.settings.project.VariantProjectDescriptor;
 import be.vbgn.gradle.buildaspects.settings.project.VariantProjectDescriptorFactory;
 import be.vbgn.gradle.buildaspects.variant.Variant;
 import be.vbgn.gradle.buildaspects.variant.VariantBuilder;
-import groovy.lang.Closure;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Function;
 import javax.inject.Inject;
-import org.gradle.api.Action;
 import org.gradle.api.Namer;
 import org.gradle.api.initialization.Settings;
 import org.gradle.api.model.ObjectFactory;
